pyproject.toml
[project]
name = "mcp-jina-supabase-rag"
version = "0.1.0"
description = "MCP server for crawling documentation and indexing to Supabase RAG using Jina AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.7.1",
    "crawl4ai>=0.6.2",
    "supabase>=2.15.1",
    "openai>=1.71.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "lxml>=5.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
